requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==19.3.0
cached-property==1.5.1
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
defusedxml==0.6.0
Django==2.2
djangorestframework==3.11.0
idna==2.9
isodate==0.6.0
py-moneyed==0.8.0
pycparser==2.20
pyOpenSSL==19.1.0
pytz==2020.1
requests==2.23.0
requests-pkcs12==1.7
requests-toolbelt==0.9.1
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
zeep==3.4.0
psycopg2==2.9.1
